% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available-packages.R
\name{available_packages}
\alias{available_packages}
\title{Download verified package indices}
\usage{
available_packages(contriburl = contrib.url(repos, type), method,
  fields = NULL, type = getOption("pkgType"), filters = NULL,
  repos = getOption("repos"), pubkey = NULL)
}
\arguments{
\item{contriburl}{URL of the contrib section of the repository}

\item{method}{Ignored, but included for compatibility with
\code{available.packages}}

\item{fields}{See \code{available.packages}}

\item{type}{The type of package}

\item{filters}{See \code{available.packages}}

\item{repos}{Repositories}

\item{pubkey}{Path or contents of public key}
}
\description{
Download verified package indices
}
