% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-releases.R
\name{get_tags}
\alias{get_tags}
\title{Retrieve a data frame of GitHub package tag (release) info}
\usage{
get_tags(repo)
}
\arguments{
\item{repo}{Repository address in the format `username/repo`}
}
\description{
Retrieve a data frame of GitHub package tag (release) info
}
