% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-packages.R
\name{install_packages}
\alias{install_packages}
\title{Install and validate packages}
\usage{
install_packages(..., pubkey = NULL)
}
\arguments{
\item{...}{Arguments passed through to \code{install.packages}}

\item{pubkey}{Public key to use}
}
\description{
Validating version of \code{install.packages}
}
