% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-release.R
\name{install_release}
\alias{install_release}
\title{Validate that the current release is signed and install it if so}
\usage{
install_release(repo)
}
\arguments{
\item{repo}{Repository address in the format \code{username/repo}}
}
\description{
Validate that the current release is signed and install it if so
}
\examples{
\dontrun{
install_release("hrbrmstr/hrbrthemes")

# fails
install_release("ironholds/rgeolocate")
}
}
