% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notary-package.R
\docType{package}
\name{notary}
\alias{notary}
\alias{notary-package}
\title{Signing and Verification of R Packages}
\description{
Signing and Verification of R Packages
}
\author{
Bob Rudis (bob@rud.is)
}
