% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-index.R
\name{package_index_prepare}
\alias{package_index_prepare}
\title{Prepare a package index}
\usage{
package_index_prepare(path, key)
}
\arguments{
\item{path}{Directory that contains a \code{PACKAGES} file and a
set of packages.}

\item{key}{Path or contents of private key}
}
\description{
Prepare a package index for use with notary; computes and adds MD5
and SHA256 hashes the index and then signs them with your private
key.
}
