% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-signature.R
\name{retrive_release_signature}
\alias{retrive_release_signature}
\title{Retrieve the GitHub signing information for the latest release of a package}
\usage{
retrive_release_signature(repo, verbose = TRUE)
}
\arguments{
\item{repo}{Repository address in the format \code{username/repo}}

\item{verbose}{if \code{TRUE} then a message will be returned with information about
the signer of the package.}
}
\description{
Check that the latest release of the package at \code{username/repo} has been signed.
}
\examples{
retrive_release_signature("hrbrmstr/hrbrthemes")

retrive_release_signature("ironholds/rgeolocate")
}
