% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-index.R
\name{sign_file}
\alias{sign_file}
\title{Sign a file with `sodium` key}
\usage{
sign_file(filename, key)
}
\arguments{
\item{filename}{path to file to sign}

\item{key}{key to use for signing}
}
\description{
Sign a file with `sodium` key
}
