% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-safe.R
\name{source_safe_sign}
\alias{source_safe_sign}
\alias{sys_source_safe_sign}
\title{Source a file with verification}
\usage{
source_safe_sign(file, ..., pubkey = NULL, verbose = FALSE)

sys_source_safe_sign(file, ..., pubkey = NULL, verbose = FALSE)
}
\arguments{
\item{file}{The file to source.  Can be a path, file:// or
https:// URL}

\item{...}{Additional parameters passed through to \code{source}
or \code{sys.source}.}

\item{pubkey}{Public key (filename or contents)}

\item{verbose}{Be verbose when downloading files?}
}
\description{
Source a file with verification
}
