% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-release.R
\name{validate_release}
\alias{validate_release}
\title{Validate that the latest release of a GitHub pacakge is}
\usage{
validate_release(repo)
}
\arguments{
\item{repo}{Repository address in the format \code{username/repo}}
}
\description{
Check that the latest release of the package at \code{username/repo} has been signed.
}
\examples{
validate_release("hrbrmstr/hrbrthemes")

validate_release("ironholds/rgeolocate")
}
